LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-back-cross-document-expected.txt



FAIL navigate event for history.back() - cross-document assert_equals: expected (object) null but got (undefined) undefined
